requirements.txt
Django==4.1.13
djangorestframework==3.14.0
